Now that's a big update

+++ SoftDesk/projects/views_nested.py
@@ -104,6 +104,9 @@ class ProjectViewSet(ModelViewSet):
 
 
 class ContributorViewSet(ModelViewSet):
+    queryset = Contributor.objects.all().select_related("project")
+    serializer_class = ContributorSerializerGet
+
     def get_permissions(self):
         """Instantiates and returns the list of
         permissions that this view requires."""
@@ -113,23 +116,24 @@ class ContributorViewSet(ModelViewSet):
             permission_classes = [ProjectAuthor]
         return [permission() for permission in permission_classes]
 
-    def retrieve(self, request, pk=None):
-        project = get_object_or_404(Project, pk=pk)
-        users = Contributor.objects.filter(project=project)
-        serializer = ContributorSerializerGet(users, many=True)
-        return Response(serializer.data)
+    def get_queryset(self, *args, **kwargs):
+        project_id = self.kwargs.get("projects_pk")
+        try:
+            project = Project.objects.get(id=project_id)
+        except Project.DoesNotExist:
+            return Response("That project does not exist")
+        return self.queryset.filter(project=project)
 
-    def create(self, request, pk=None):
-        # project = Project.objects.get(pk=pk)
-        project = get_object_or_404(Project, pk=pk)
+    def create(self, request, *args, **kwargs):
+        project_pk = self.kwargs.get("projects_pk")
 
         copied_data = request.data.copy()
-        copied_data["project"] = project.id
+        copied_data["project"] = project_pk
         copied_data["role"] = "contributor"
 
         try:
             contributor = Contributor.objects.get(
-                user=copied_data["user"], project=project.id
+                user=copied_data["user"], project=project_pk
             )
             return Response(
                 f"[{contributor.user}] (ID: {contributor.user} is already part of the project.",
@@ -152,8 +156,12 @@ class ContributorViewSet(ModelViewSet):
             except User.DoesNotExist:
                 return Response(f"That user does not exist.")
 
-    def destroy(self, request, pk, contributor_pk):
-        contributor = get_object_or_404(Contributor, user=contributor_pk, project=pk)
+    def destroy(self, request, *args, **kwargs):
+        project_pk = self.kwargs.get("projects_pk")
+        contributor_pk = self.kwargs.get("pk")
+        contributor = get_object_or_404(
+            Contributor, user=contributor_pk, project=project_pk
+        )
 
         if contributor.role == "author":
             return Response(
@@ -169,8 +177,8 @@ class ContributorViewSet(ModelViewSet):
 
 
 class IssueViewSet(ModelViewSet):
-    # queryset = Issue.objects.all().select_related("project")
-    # serializer_class = IssueSerializerGet
+    queryset = Issue.objects.all().select_related("project")
+    serializer_class = IssueSerializerGet
 
     def get_permissions(self):
         """Instantiates and returns the list of
@@ -181,49 +189,36 @@ class IssueViewSet(ModelViewSet):
             permission_classes = [IssueAuthor]
         return [permission() for permission in permission_classes]
 
-    def list(self, request, pk=None):
-        project = Project.objects.get(pk=pk)
-        issues = Issue.objects.filter(project=project)
-        serializer = IssueSerializerGet(issues, many=True)
+    def get_queryset(self, *args, **kwargs):
+        project_id = self.kwargs.get("projects_pk")
+        try:
+            project = Project.objects.get(id=project_id)
+        except Project.DoesNotExist:
+            return Response("That project does not exist")
+        return self.queryset.get(project=project)
 
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-    def create(self, request, pk=None):
-        project = Project.objects.get(pk=pk)
-        project_id = project.id
+    def create(self, request, *args, **kwargs):
+        project_pk = self.kwargs.get("projects_pk")
 
         copied_data = request.data.copy()
-        copied_data["project"] = project_id
+        copied_data["project"] = project_pk
         copied_data["author"] = request.user.id
 
-        tags = ["bug", "improvement", "task"]
-        copied_data["tag"] = copied_data["tag"].casefold()
-        if copied_data["tag"] not in tags:
-            return Response("Available tags : bug, improvement, task")
-
-        priorities = ["low", "moderate", "high"]
-        copied_data["priority"] = copied_data["priority"].casefold()
-        if copied_data["priority"] not in priorities:
-            return Response("Available priorities : low, moderate, high")
-
-        status_list = ["to_do", "in_progress", "completed"]
-        copied_data["status"] = copied_data["status"].casefold()
-        if copied_data["status"] not in status_list:
-            return Response("Available priorities : to_do, in_progress, completed")
+        print(f"\nrequest.user.id {request.user.id}\n")
 
         if copied_data["assignee"] == "":
             copied_data["assignee"] = request.user.id
 
         try:
             contributor = get_object_or_404(
-                Contributor, user=copied_data["assignee"], project=project_id
+                Contributor, user=copied_data["assignee"], project=project_pk
             )
             copied_data["assignee"] = contributor.id
         except Exception:
             pass
 
         try:
-            Contributor.objects.get(id=copied_data["assignee"], project=project_id)
+            Contributor.objects.get(id=copied_data["assignee"], project=project_pk)
             serializer = IssueSerializer(data=copied_data)
             if serializer.is_valid(raise_exception=True):
                 serializer.save()
@@ -235,61 +230,30 @@ class IssueViewSet(ModelViewSet):
                 status=status.HTTP_400_BAD_REQUEST,
             )
 
-    def update(self, request, pk, issue_pk):
-        project = get_object_or_404(Project, id=pk)
-        project_id = project.id
+    def update(self, request, *args, **kwargs):
+        project_pk = self.kwargs.get("projects_pk")
+        issue_pk = self.kwargs.get("pk")
+
         issue = get_object_or_404(Issue, id=issue_pk)
         copied_data = request.data.copy()
+        copied_data["tag"] = issue.tag
         copied_data["project"] = issue.project.id
         copied_data["author"] = issue.author.id
+        # copied_data["assignee"] = issue.assignee.id
 
-        tags = ["bug", "improvement", "task"]
-        copied_data["tag"] = copied_data["tag"].casefold()
-        if copied_data["tag"] not in tags:
-            return Response("Available tags : bug, improvement, task")
-
-        priorities = ["low", "moderate", "high"]
-        copied_data["priority"] = copied_data["priority"].casefold()
-        if copied_data["priority"] not in priorities:
-            return Response("Available priorities : low, moderate, high")
-
-        status_list = ["to_do", "in_progress", "completed"]
-        copied_data["status"] = copied_data["status"].casefold()
-        if copied_data["status"] not in status_list:
-            return Response("Available priorities : to_do, in_progress, completed")
-
-        # ----------- Checking the "user" input
-        project_contributors = Contributor.objects.filter(project=pk)
-        contrib_users_list = []
-        for contrib in project_contributors:
-            contrib_users_list.append(contrib.user.username)
-
-        user_input = copied_data["assignee"]
-        if user_input.isdigit():
-            pass
-        elif user_input in contrib_users_list:
-            picked_user = User.objects.get(username=user_input)
-            copied_data["assignee"] = picked_user.id
-        else:
-            print(
-                f"Blank input or user [{user_input}] does not exist, authenticated user [{request.user}] selected as assignee."
-            )
+        if copied_data["assignee"] == "":
             copied_data["assignee"] = request.user.id
-        # ---------------------------------------------
-
-        # if copied_data["assignee"] == "":
-        #     copied_data["assignee"] = request.user.id
 
         try:
             contributor = get_object_or_404(
-                Contributor, user=copied_data["assignee"], project=project_id
+                Contributor, user=copied_data["assignee"], project=project_pk
             )
             copied_data["assignee"] = contributor.id
         except Exception:
             pass
 
         try:
-            Contributor.objects.get(id=copied_data["assignee"], project=project_id)
+            Contributor.objects.get(id=copied_data["assignee"], project=project_pk)
             serializer = IssueSerializer(issue, data=copied_data)
             if serializer.is_valid(raise_exception=True):
                 serializer.save()
@@ -301,9 +265,9 @@ class IssueViewSet(ModelViewSet):
                 status=status.HTTP_400_BAD_REQUEST,
             )
 
-    def destroy(self, request, pk, issue_pk):
+    def destroy(self, request, *args, **kwargs):
+        issue_pk = self.kwargs.get("pk")
         issue = get_object_or_404(Issue, id=issue_pk)
-        # issue = Issue.objects.get(id=issue_pk)
 
         issue.delete()
         return Response(
@@ -328,26 +292,33 @@ class CommentViewSet(ModelViewSet):
             permission_classes = [CommentAuthor]
         return [permission() for permission in permission_classes]
 
-    def list(self, request, pk, issue_pk):
-        issue = get_object_or_404(Issue, id=issue_pk)
+    def list(self, request, projects_pk, pk):
+        # issue = Issue.objects.get(id=issue_pk)
+        issue = get_object_or_404(Issue, id=pk)
         comments = Comment.objects.filter(issue_id=issue)
         serializer = CommentSerializerGet(comments, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def create(self, request, pk, issue_pk):
-        issue = get_object_or_404(Issue, id=issue_pk)
-
+    def create(self, request, projects_pk, pk):
+        # project = get_object_or_404(Project, id=projects_pk)
+        issue = get_object_or_404(Issue, id=pk)
+        print(f"\nISSUE ID {issue.id}\n")
         copied_data = request.data.copy()
         copied_data["issue_id"] = issue.id
         copied_data["author"] = request.user.id
 
         serializer = CommentSerializer(data=copied_data)
+        print(f"\nSERiALIZER {serializer}\n")
         if serializer.is_valid():
+            print("\nVALIDATOR OK\n")
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
+        print("\nVALIDATOR NOOOOOOOT OK\n")
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def update(self, request, pk, issue_pk, comment_pk):
+    def update(self, request, projects_pk, pk, comment_pk):
+        # project = get_object_or_404(Project, id=projects_pk)
+        # issue = get_object_or_404(Issue, id=pk)
         comment = get_object_or_404(Comment, id=comment_pk)
         copied_data = request.data.copy()
         copied_data["issue_id"] = comment.issue_id.id
@@ -359,15 +330,21 @@ class CommentViewSet(ModelViewSet):
             return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def retrieve(self, request, pk, issue_pk, comment_pk):
+    def retrieve(self, request, projects_pk, pk, comment_pk):
         comment = get_object_or_404(Comment, id=comment_pk)
         serializer = CommentSerializerGet(comment)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def destroy(self, request, pk, issue_pk, comment_pk):
+    def destroy(self, request, projects_pk, pk, comment_pk):
         comment = get_object_or_404(Comment, id=comment_pk)
 
-        comment.delete()
+        if request.user == comment.author:
+            comment.delete()
+            return Response(
+                "The comment was delete successfully.", status=status.HTTP_202_ACCEPTED
+            )
+
         return Response(
-            "The comment was deleted successfully.", status=status.HTTP_202_ACCEPTED
+            "Only the author of the comment can delete that comment.",
+            status=status.HTTP_400_BAD_REQUEST,
         )
